works better with resizer
the Jquery will cause the tabbed chatlog to be aware if the sidebar has resized, either via system prefs, or a resizing module.
The resize takes effect the next time a notification appears.   Ideally it would move the notifications in real-time, but people usually don't adjust the sidebar size repeatedly.

Currently the notifications appear at a fixed position regardless of the sidebar width and this can confuse players to click on the wrong Tab to read the message which.